refactor(runtime): split the entry point by concern
Splits the 875-line packages/runtime/src/index.ts (mixing ~7 unrelated
concerns per refactor-plan.md §1.2-F/§3.3) into directive.ts, field.ts,
interface.ts, object.ts, scalar.ts, resolver.ts, and apis.ts, with
index.ts reduced to a barrel re-exporting the exact same 34 public names
-- the type-level tests from the previous commit pass unmodified against
the new layout, confirming the public surface is byte-compatible.

Also documents Decision D5 (runtime half): defineQuery/defineMutation/
defineField/defineSubscription now spell out in their TSDoc that explicit
type arguments are required, since the CLI reads the syntactic type
arguments off the call expression rather than relying on inference.

packages/cli/src/core/metadata-contract.test.ts pinned the marker-string
contract to a single runtime/src/index.ts path; it now scans every
non-test .ts file under runtime/src so the contract keeps holding now
that the marker literals are spread across the new files.
